IdGenerator 인터페이스 제네릭 추가 generate 메소드 반환 타입 수정

+++ app/src/main/java/com/codesoom/assignment/services/TaskService.java
@@ -33,7 +33,7 @@ public class TaskService implements TaskServiceInterface{
 
     @Override
     public void createTask(TaskDTO taskDTO){
-        Long nextId = gen.generate();
+        Long nextId = (Long) gen.generate();
         tasks.put(nextId , mapper.toNewTask(nextId , taskDTO));
     }
